fix migration order

+++ db/migrate/20191122111708_create_words.rb
@@ -1,13 +1,13 @@
 class CreateWords < ActiveRecord::Migration[6.0]
   def change
     create_table :words do |t|
-      t.numeric :rank, default: 0
+      t.integer :rank, default: 0
       t.string :pos, null: false
       t.string :en, null: true
       t.string :pt, null: true
       t.string :es, null: true
       t.boolean :status, default: true
-      t.numeric :level, default: 0
+      t.integer :level, default: 0
 
       t.timestamps
     end
